requirements.txt
# Django Core
Django>=4.2,<5.0
django-environ>=0.11.0

# Database
dj-database-url>=2.0.0
psycopg2-binary>=2.9.0

# Django REST Framework
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# API Documentation
drf-spectacular>=0.27.0
drf-spectacular-sidecar>=2024.1.1

# CORS
django-cors-headers>=4.3.0

# Static Files
whitenoise>=6.6.0

# Health Checks
django-health-check>=3.18.0

# Background Jobs
django-rq>=2.10.0
redis>=5.0.0

# Security
django-axes>=6.1.0

# Logging
python-json-logger>=2.0.0

# Monitoring (optional)
sentry-sdk>=1.40.0

# WSGI Server
gunicorn>=21.2.0

# Utils
python-decouple>=3.8
requests>=2.31.0
